move unit tests to new file
